Add About Us, Update Refute UIS Response and Change Title to Respond to Pushback (#204)
* Update Readme

* Remove Prisma previewFeatures

Reponding to the following prisma warning messages:

```
Preview feature "clientExtensions" is deprecated. The functionality can be used without specifying it as a preview feature.
Preview feature "orderByNulls" is deprecated. The functionality can be used without specifying it as a preview feature.
```

* Add About Us Section

* Change the Timeline Background to Blue to match background color pattern

* Pad timelineSection in X and Y direction

* Remove Sample Section

* Correct Section Name Spelling

* Continue to Update Component Spelling

* Remove timeline overflow scroll bar

the bar is not needed as the component class ensure there will be no
overflow

* Implement Claim Response Cards

* Add subheader-1 with graph-ff-condensed fonts


---------

Co-authored-by: kennykos <“gkosmacher01@gmail.com”>